section 8 (partial): split hfs_fsck Phase 4 (bitmap) into submodule
Extract check_extents_and_bitmap, collect_overflow_extents,
collect_bad_block_extents, mark_extents, repair_extents,
repair_mdb_counts_and_bitmap into src/fs/hfs_fsck/bitmap.rs (698 lines).
The local MAX_BITMAP_MISMATCHES const and hfs_debug_issue helper move
with them since only Phase 4 uses them. CatalogKey, CatalogEntry, and
collect_catalog_entries bumped to pub(super) so the submodule can call
into the catalog parser still living in mod.rs.

hfs_fsck/mod.rs drops from 5031 to 4353 lines. All 61 fsck tests pass.
