+++ docs/codecleanup.md
@@ -177,8 +177,8 @@ Conventions:
 - [ ] Catalog walking / record-range helpers shared with HFS+ (§1).
 - [ ] Audit debug `eprintln!` instances flagged in `hfs.rs:3068–3124` and replace with `log::debug!`.
 - [~] `hfs_fsck.rs` split started: extents-overflow code moved to its own submodule.
-  - **Done:** `src/fs/hfs_fsck.rs` is now `src/fs/hfs_fsck/mod.rs` (5031 lines, was 6588) plus two submodules: `extents.rs` (1337 lines, the 17 `*_extents_*` functions for the extents-overflow B-tree — `check_extents_btree_structure` / `repair_extents_btree_structure` + helpers + tests) and `mdb.rs` (260 lines, Phase 1 MDB sanity — `validate_hfs_name`, `check_alternate_mdb`, `check_embedded_hfs_plus`, `check_mdb`). Helpers `HfsFsckCode`, `HFS_MAX_NRECS`, `hfs_issue`, `extract_child_pointers`, `record_key` bumped to `pub(super)` so submodules can reach them. All 61 fsck tests still pass under their new module paths.
-  - **Not yet done:** further per-phase splits (catalog / btree / bitmap). The remaining mod.rs has tighter cross-references between those phases (catalog and B-tree share many helpers like `compare_catalog_keys`, `walk_leaf_chain`, `repair_btree_structure_leaves_only`); each additional split needs a careful pass over `pub(super)` visibility. Worth doing iteratively.
+  - **Done:** `src/fs/hfs_fsck.rs` is now `src/fs/hfs_fsck/mod.rs` (4353 lines, was 6588) plus three submodules: `extents.rs` (1334 lines, the 17 `*_extents_*` functions for the extents-overflow B-tree — `check_extents_btree_structure` / `repair_extents_btree_structure` + helpers + tests), `mdb.rs` (260 lines, Phase 1 MDB sanity — `validate_hfs_name`, `check_alternate_mdb`, `check_embedded_hfs_plus`, `check_mdb`), and `bitmap.rs` (698 lines, Phase 4 extent/allocation-bitmap cross-check + repairs — `check_extents_and_bitmap`, `collect_overflow_extents`, `mark_extents`, `repair_extents`, `repair_mdb_counts_and_bitmap`, plus the local `MAX_BITMAP_MISMATCHES` const and `hfs_debug_issue` helper). Helpers `HfsFsckCode`, `HFS_MAX_NRECS`, `hfs_issue`, `extract_child_pointers`, `record_key`, `CatalogKey`, `CatalogEntry`, `collect_catalog_entries` bumped to `pub(super)` so submodules can reach them. All 61 fsck tests still pass under their new module paths.
+  - **Not yet done:** further per-phase splits (catalog / btree). The remaining mod.rs has tighter cross-references between those two phases (catalog and B-tree share helpers like `compare_catalog_keys`, `walk_leaf_chain`, `repair_btree_structure_leaves_only`); each additional split needs a careful pass over `pub(super)` visibility. Recommended order: btree first (the hub) then catalog. Worth doing iteratively.
 
 ### HFS+ (`src/fs/hfsplus.rs`, 2607)
 - [ ] Migrate to `btree_split_leaf_with_insert` (§1).
